Bugfixes for icx
GEOSgcm (and thus GEOSadas) will soon move to use Intel ifort 2021.13.
With this comes the `icx` C compiler. Testing showed that some of the
CMake was a bit loose and, for Intel, was applying `-extend-source` to
not just Fortran files, but also *C* files. This is not a C flag (for
`icx`) so we need to update the CMake to fix that.